Use loopback-swagger to generate swagger.json
