+++ package.json
@@ -34,6 +34,7 @@
     "cors": "^2.7.1",
     "debug": "^2.2.0",
     "lodash": "^3.10.0",
+    "loopback-swagger": "^2.1.0",
     "strong-swagger-ui": "^21.0.0"
   }
 }
